services/courses-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
prometheus-client==0.20.0
psycopg2-binary==2.9.9